fix: install tools and rename folder back to tools
