+++ backend/src/tools/summaries.py
