+++ backend/src/tools/text_files.py
